requirements.txt
fastapi
uvicorn
py-webauthn